Move navigation package out of ui package
