+++ app/src/main/kotlin/uk/govuk/app/ui/GovUkApp.kt
@@ -27,14 +27,14 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import uk.govuk.app.design.ui.theme.GovUkTheme
-import uk.govuk.app.home.ui.navigation.HOME_GRAPH_ROUTE
-import uk.govuk.app.home.ui.navigation.homeGraph
+import uk.govuk.app.home.navigation.HOME_GRAPH_ROUTE
+import uk.govuk.app.home.navigation.homeGraph
 import uk.govuk.app.launch.AppLaunchState
 import uk.govuk.app.launch.AppLaunchViewModel
 import uk.govuk.app.navigation.TopLevelDestination
-import uk.govuk.app.onboarding.ui.navigation.ONBOARDING_GRAPH_ROUTE
-import uk.govuk.app.onboarding.ui.navigation.onboardingGraph
-import uk.govuk.app.settings.ui.navigation.settingsGraph
+import uk.govuk.app.onboarding.navigation.ONBOARDING_GRAPH_ROUTE
+import uk.govuk.app.onboarding.navigation.onboardingGraph
+import uk.govuk.app.settings.navigation.settingsGraph
 
 @Composable
 fun GovUkApp() {
